:pencil2: feat: 정복자 선정 기준을 추가로 세우고, 반환값으로 자격 여부와 총점을 반환하도록 수정

+++ backend/game-service/src/main/java/com/popplar/gameservice/entity/Game.java
@@ -36,10 +36,10 @@ public class Game extends BaseEntity {
     private Long hotPlaceId;
 
     @NotNull
-    private int points;
+    private double points;
 
     @Builder
-    public Game(GameType type, Long memberId, Long hotPlaceId, int points) {
+    public Game(GameType type, Long memberId, Long hotPlaceId, double points) {
         this.type = type;
         this.memberId = memberId;
         this.hotPlaceId = hotPlaceId;
